+++ src/components/Explorer/Result/ResultHelpers.jsx
